LPS-55590 Copy EditOrganizationAction to portlet, no logic changes
